Add ManagedStateMachine component
